Resolving numpy overflow warning

+++ src/bumpfunction.py
@@ -82,7 +82,7 @@ class BumpFunction():
 		self.output=[]
 		for i in x:
 			numerator1 = np.square(i) + 1
-			numerator2 = np.exp(4*i/(np.square(i)-1))
+			numerator2 = np.exp(4*i/(np.square(i)-1), dtype=np.float128)
 			denominator1 = np.square(i) - 1
 			denominator2 = 1 + numerator2
 			final = (numerator1 * numerator2) /  np.square(denominator1 * denominator2)
